Changes to remove compiler warnings

+++ test/transpose_inplace.cpp
@@ -55,7 +55,7 @@ void transposeip_test(af::dim4 dims)
     ASSERT_EQ(AF_SUCCESS, af_get_data_ptr((void*)trsData, inArray));
 
     dim_type nElems = dims.elements();
-    for (size_t elIter = 0; elIter < nElems; ++elIter) {
+    for (int elIter = 0; elIter < nElems; ++elIter) {
         ASSERT_EQ(trsData[elIter] , outData[elIter])<< "at: " << elIter << std::endl;
     }
 
@@ -98,7 +98,7 @@ void transposeInplaceCPPTest()
     input.host((void*)trsData);
 
     dim_type nElems = dims.elements();
-    for (size_t elIter = 0; elIter < nElems; ++elIter) {
+    for (int elIter = 0; elIter < nElems; ++elIter) {
         ASSERT_EQ(trsData[elIter], outData[elIter])<< "at: " << elIter << std::endl;
     }
